+++ elmyra.ip.access.epo/CHANGES.rst
@@ -20,7 +20,8 @@ development
 - auth: turn on open access to “kindcodes” api again
 - middleware: improve sdp backend, fetch multiple resources (xml:pretty,json:pretty,png,pdf)
 - DEPATISnet adapter: improve error detection on upstream result decoding errors
-- DEPATISnet adapter: fix XLS decoding error, upstream added new status line to first row, e.g. "Search query: pn=(EP666666) Status: 25.09.2015"
+- DEPATISnet adapter: fix XLS decoding error, upstream added new status line to first row
+  e.g. "Search query: pn=(EP666666) Status: 25.09.2015"
 
 0.100.0
 =======